Code correction

+++ OnboardingGame/OnboardingGame/OnboardingGame/Data/Database.cs
@@ -1,5 +1,6 @@
 ﻿using SQLite;
 using SQLiteNetExtensions;
+using SQLiteNetExtensionsAsync.Extensions;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -19,8 +20,6 @@ namespace OnboardingGame.Data
             _database.CreateTableAsync<TaskItem>().Wait();
             _database.CreateTableAsync<ToDoList>().Wait();
             _database.CreateTableAsync<PlayerProfile>().Wait();
-            _database.CreateTableAsync<Catagory>().Wait();
-            _database.CreateTableAsync<Achievement>().Wait();
         }
 
         public void DeleteDatabase() {
@@ -37,52 +36,92 @@ namespace OnboardingGame.Data
         }
         public Task<TaskItem> GetTaskItem(int id)
         {
-            if (id >= 0 && _database.Table<ToDoList>().ToListAsync().Result.Count > id)
-            {
-                return _database.Table<TaskItem>().Where(i => i.ID == id).FirstOrDefaultAsync();
-            }
-            return null;
+            return _database.Table<TaskItem>().Where(i => i.ID == id).FirstOrDefaultAsync();    
         }
 
         public Task<int> SaveItemAsync(TaskItem item)
         {
-            if (item.ID != 0)
+            if (!(item is null))
             {
-                return _database.UpdateAsync(item);
+                if (item.ID != 0)
+                {
+                    return _database.UpdateAsync(item);
+                }
+                else
+                {
+                    return _database.InsertAsync(item);
+                }
             }
-            else
+            return null;
+        }
+        public void SaveItemAsync(List<TaskItem> items) {
+            if (!(items is null))
             {
-                return _database.InsertAsync(item);
+                for (int i = 0; i < items.Count; i++)
+                {
+                    SaveItemAsync(items[i]).Wait();
+                }
             }
         }
         public Task<int> DeleteItemAsync(TaskItem item)
         {
-            return _database.DeleteAsync(item);
+            if (!(item is null))
+            {
+                return _database.DeleteAsync(item);
+            }
+            return null;
         }
 
         //ToDoList__________________________________________________________________
         public Task<List<ToDoList>> GetToDoListAsync()
         {
-            return _database.Table<ToDoList>().ToListAsync();
+            return _database.GetAllWithChildrenAsync<ToDoList>();
+            //return _database.Table<ToDoList>().ToListAsync();
         }
         public Task<ToDoList> GetToDoListAsync(int id)
         {
-            return _database.Table<ToDoList>().Where(i => i.ID == id).FirstOrDefaultAsync();
+            bool exist = _database.Table<ToDoList>().Where(i => i.ID == id).FirstOrDefaultAsync().Result != null;
+
+            if (id > 0 && exist)
+            {
+                return _database.GetWithChildrenAsync<ToDoList>(id);
+            }
+            return null;
+            //return _database.Table<ToDoList>().Where(i => i.ID == id).FirstOrDefaultAsync();
         }
         public Task<int> SaveListAsync(ToDoList item)
         {
-            if (item.ID != 0)
+            if (!(item is null))
             {
-                return _database.UpdateAsync(item);
+                if (item.ID != 0)
+                {
+                    _database.UpdateWithChildrenAsync(item).Wait();
+                    //return 1;
+                }
+                else
+                {
+                    _database.InsertWithChildrenAsync(item).Wait();
+                    //return 1;
+                }
             }
-            else
+            return null;
+        }
+        public void SaveListAsync(List<ToDoList> items) {
+            if (!(items is null))
             {
-                return _database.InsertAsync(item);
+                for (int i = 0; i < items.Count; i++) {
+                    SaveListAsync(items[i]);
+                }
             }
         }
+
         public Task<int> DeleteItemAsync(ToDoList item)
         {
-            return _database.DeleteAsync(item);
+            if (!(item is null))
+            {
+                return _database.DeleteAsync(item);
+            }
+            return null;
         }
 
         //PlayerProfile_____________________________________________________________
@@ -90,56 +129,26 @@ namespace OnboardingGame.Data
         {
             return _database.Table<PlayerProfile>().FirstOrDefaultAsync();
         }
+        public Task<PlayerProfile> GetPlayerProfile(int id) 
+        {
+            return _database.Table<PlayerProfile>().Where(i => i.ID == id).FirstOrDefaultAsync();
+        }
         public Task<int> SavePlayerAsync(PlayerProfile item) {
-            if (item.ID != 0)
+            if (!(item is null))
             {
-                return _database.UpdateAsync(item);
+                if (item.ID != 0)
+                {
+                    return _database.UpdateAsync(item);
+                }
+                else
+                {
+                    return _database.InsertAsync(item);
+                }
             }
-            else
-            {
-                return _database.InsertAsync(item);
-            }
+            return null;
         }
         public Task<int> DeletePlayerAsync() {
             return _database.DeleteAllAsync<PlayerProfile>();
-        }
-
-        //Catagories________________________________________________________________
-        public Task<List<Catagory>> GetCatagories() {
-            return _database.Table<Catagory>().ToListAsync();
-        }
-        public Task<Catagory> GetCatagory(int id) {
-            return _database.Table<Catagory>().Where(i => i.ID == id).FirstOrDefaultAsync();
-        }
-        public Task<int> SaveCatagoryAsync(Catagory item)
-        {
-            return _database.InsertAsync(item);
-        }
-        public Task<int> DeleteCatagoryAsync(Catagory item)
-        {
-            return _database.DeleteAsync(item);
-        }
-
-        //Achievements______________________________________________________________
-        public Task<List<Achievement>> GetAchievement() {
-            return _database.Table<Achievement>().ToListAsync();
-        }
-        public Task<Achievement> GetAchievement(int id) {
-            return _database.Table<Achievement>().Where(i => i.ID == id).FirstOrDefaultAsync();
-        }
-        public Task<int> SaveAchievement(Achievement item) {
-            if (item.ID != 0)
-            {
-                return _database.UpdateAsync(item);
-            }
-            else
-            {
-                return _database.InsertAsync(item);
-            }
-        }
-        public Task<int> DeleteAchievement(Achievement item) {
-
-            return _database.DeleteAsync(item);
         }
 
         //Return a list of TaskItems who's ListID matches that of the given id 
@@ -150,15 +159,12 @@ namespace OnboardingGame.Data
         public Task<ToDoList> GetLatestSavedList() {
             return _database.Table<ToDoList>().ElementAtAsync(_database.Table<ToDoList>().CountAsync().Result - 1);
         }
-        public Task<int> GetAllDoneTasks(int id)
+        public Task<List<TaskItem>> GetAllDoneTasks(int id)
         {
-            return _database.Table<TaskItem>().Where(i => i.ListID == id && i.Status >= 1).CountAsync();
+            return _database.Table<TaskItem>().Where(i => i.ListID == id && i.status >= 1).ToListAsync();
         }
-        public Task<int> GetAllDoneTasks() {
-            return _database.Table<TaskItem>().Where(i => i.Status >= 1).CountAsync();
-        }
-        public Task<List<Achievement>> GetAchievementByType(App.AchievementType t) {
-            return _database.Table<Achievement>().Where(i => i.AchievementType == t && i.Status == false).ToListAsync();
+        public Task<List<TaskItem>> GetAllDoneTasks() {
+            return _database.Table<TaskItem>().Where(i => i.status >= 1).ToListAsync();
         }
     }
 }
